Typecasted int to pointer

+++ src/AvlString.h
@@ -5,7 +5,7 @@
 #include "Node.h"
 
 #define avlAddString(r,n)     avlAdd((Node **)(r),(Node *)(n),(Compare)(StringCompare))
-#define avlRemoveString(r,n)  avlRemove((Node **)(r),(int )(n),(Compare)(StringCompare))
+#define avlRemoveString(r,n)  avlRemove((Node **)(r),(intptr_t)(n),(Compare)(StringCompare))
 
 int StringCompare(char* nodeData,StringNode *refNode);
